FIXED: Problems when saving and updating promotions
FIXED: Conflicts between varchar and enum type in the database. A cast was added in the database. Also, the createdAt and updatedAt fields where updated to not register a null in the database as it automatically takes the current timestamp and stores it.

FIXED: Wrong class was being saved when an update was made to a promotion, creating duplicates