REFACTORED: Message microservice configuration
The configuration, which was moved into config-server for the sake of consistency has been brought back to its original application.properties file and the fetching from config-server has been removed

This is because of a problem related to a bean which was necessary for the mail to even boot up. Said bean was created in run time, it didn't wait for the configurations from config-server to load

There was no way of delaying the creation of the bean, of creating a manual bean and then being overriden. NOTHING

Maybe there's another way but simply moving te configuratio back was the simplest solution